--- neurobench/examples/obj_detection/model_data/model_parameters
@@ -1,1 +1,0 @@
-This folder should contain trained model parameters, if applicable.
